requirements.txt
numpy
pandas
scikit-learn
datetime
dvc
dvc-s3
mlflow
pymongo
python-dotenv
joblib
dagshub